+++ src/Installer/SymfonyInstaller.php
@@ -53,7 +53,7 @@ final class SymfonyInstaller extends AbstractServiceInstaller implements NeedsDa
         }
 
         if (($answers['database'] ?? null) !== null) {
-            $expression->callMethod('withDatabaseService', [Ast::var((string) $answers['database'])]);
+            $expression->callMethod('link', [Ast::var((string) $answers['database'])]);
         }
     }
 
@@ -70,7 +70,7 @@ final class SymfonyInstaller extends AbstractServiceInstaller implements NeedsDa
         }
 
         if (($answers['database_instance'] ?? null) instanceof DatabaseServiceInterface) {
-            $service->withDatabaseService($answers['database_instance']);
+            $service->link($answers['database_instance']);
         }
 
         return $service;